Makefile: Use -Wno-missing-field-initializers on older GCC
When compiling under CentOS 7 with GCC 4.8.5 we got the following

example-main.c:51:2: warning: missing initializer for field ‘color’ of ‘const struct tc_coloris’ [-Wmissing-field-initializers]
  {}
  ^
In file included from example-main.c:27:0:
textus_coloris.h:47:14: note: ‘color’ declared here
  const char *color;
              ^

In the examples we have an array of structs defining the colours, this
array is terminated by '{}'.

While that is not strictly allowed in C (it is in C++), it works in both
GCC >= 5.1.1 and Clang. This is used extensively in the Linux kernel and
I think it just looks better than having to fill out the entry with
0/NULLs.

For older GCC we get the above warning which can be suppressed with
-Wno-missing-field-initializers, so add a GCC version check and add that
option for GCC < 5.1.1

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>